requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg[binary]
alembic
python-dotenv
pydantic
pydantic-settings